backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai
pydantic-settings
python-dotenv
json-repair
python-multipart
